fixes JOO-43 Using EXML target class elements does not trigger target class import http://jangaroo.myjetbrains.com/youtrack/issue/JOO-43
In the generated target class, for each usage of another class, in addition
to an import of the config class, also adding an import of the target
class. That does no harm concerning public API violations, since it is
not forbidden to import a non-public-API class as long as it is not
actually used by code.