first take at supporting module loading of native scripts
You can now add an "amd" attribute to the [Native] annotation,
and using the AS API leads to a module dependency on the
value of the "amd" attribute is added. For example:

package foo {
[Native(amd="foo/my-native-class")]
public class MyNativeClass { ... }
}

When using foo.MyNativeClass, the client class gets a dependency
on "foo/my-native-class" and (like before) expects foo.MyNativeClass
to be globally available afterwards.