fixed access to static in-scope members of non-imported super-super class
This occurred in JooUnit, where TestCase inherits from Assert so that
TestCase subclasses can use assertTrue() and the like without qualifier
or import.
The mean thing is that when making Assert.assertTrue() explicit through
a synthesized DotExpr, it would have to be imported, or the Ide (here:
"Assert") cannot be resolved later when generating code.
Solution: inject the resolved declaration into the Ide, so it does not even
try to resolve itself through imports.