Support for MXML component libraries
As in Flex, you can now define custom component libraries,
using a manifest.xml to map element names to classes and
compiler options to given the library a namespace URI.
Also as in Flex, the component library is compiled into a
catalog.xml file that is included in the JAR module artifact.
Unfortunately, IDEA does not find catalog.xml files inside JARs,
but only inside SWCs, so there we are not finished yet.