Prepare to make MXML component libraries work in IDEA
Since we do not build SWCs, we have to bundle the manifest.xml(s)
and the namespace-to-manifest mapping inside the JAR.
We chose to use the Flex compiler configuration format for the
namespace-to-manifest mapping.
The Jangaroo IDEA Plugin can then scan all dependent JARs for
a config.xml and thus find all custom component namespaces.
These can then be added to the IDEA module's compiler configuration.