fix statics test, which broke with commit 2023fee385ba2f185efa0ab030a132c6a2492582
Solution is still not optimal, since we should really distinguish between static and non-static
members through dedicated types for both. In that sense, a class defines two types: one for
its non-static members, including inheritance, and one for its static members, where static
members of superclasses are only in scope of the subclass, but not inherited. Currently, we
cannot express the latter type, so we have to live with this workaround.