fixes JOO-70 for-(each-)in loop over Array may enumerate additional elements and/or produce wrong order http://jangaroo.myjetbrains.com/youtrack/issue/JOO-70
for-(each-)in-loops over Array-typed expression are now
rewritten to indexed loops, using an auxiliary variable.
Not that Array expression type detection only works for
Array literals and (qualified) identifiers, not for more complex
expressions like method calls. As a workaround, you have to
introduce a local variable of type Array.