fixes JOO-73 EXML validation may fail even in validationMode "warn" http://jangaroo.myjetbrains.com/youtrack/issue/JOO-73
Re-declaring config options in a subclass (even with the same
type) led to an invalid XSD being generated, which in turn,
instead of logging a warning or an error, threw an exception.
Fixed filtering of duplicate config options: even after fixing
a bug that caused fully qualified config attribute types parsed
from an ActionScript file to be incorrect, there still may be
inconsistencies in the type representation--fully qualified
versus imported--that lead to non-equal types, so for the
time being, ConfigAttributes (to be precise: all TypedFields)
no longer have to have the same type to be considered
equal, but only the same name.
We now use these "cleaned up" config attributes not only
when generating the XSD, but also In the config class template,
so that no longer property accessors are generated that are
red without an "override" keyword.
Even if something still slips through that leads to an invalid
XSD, such errors are no longer re-thrown, but logged as
configured (warning or error).
All EXML validation messages now start with "EXML validation:"
so that they can easily be extracted by tools.