Support method annotations again
Introduced runtime metadata in a format that now covers everything
possible in "real" AS3:
* multiple annotations of the same name at the same element
  (e.g. multiple [Event] annotations at a class)
* multiple annotation arguments with the same name

For Jangaroo 2 backwards-compatibility, JooClassDeclaration and the
re-introduced class MemberDeclaration provide metadata in the
old format, where the annotation name and the argument names are
keys in a map (Object), so there can only be one value.
An API for full access to metadata is to follow later.
So far, only runtime annotations on classes and non-static methods
and fields are kept.
The following well-known annotations are not kept at runtime (this
will be configurable): "Embed", "Native", "Accessor"