Jangaroo-3-specific bug: "this" is untyped in field initializers Consequences: * own method references (this.getFoo without invocation, i.e. not this.getFoo()) are not bound in field initializers * private member names are not rewritten when used in field initializers
Evaluation:
Jangaroo 3 rewrites qualified identifiers like this.getFoo to dot expressions <this>.<getFoo>
and tries to resolve the left-hand side expression. This fails in initializers, because while they
are in instance scope, the typed "this" is only declared as an implicit parameter in non-static
methods.
Fix:
Changed that so that typed "this" is declared in instance scope and no longer an implicit
parameter.