fixes JOO-77 http://jangaroo.myjetbrains.com/youtrack/issue/JOO-77 Imported class not generated in time during EXML compilation
Implemented a new strategy when to generate AS classes from EXML.
ConfigClassRegistry now takes over several of the responsibilities of Exmlc.
Exmlc merely delegates to its ConfigClassRegistry.
Instead of scanning all EXML files using an XML parser, ConfigClassRegistry
now starts with scanning only the file system for all EXML sources, creating
instances of the new helper class ExmlSourceFile. Each ExmlSourceFile instance
keeps track of one to-be-compiled EXML source file, which is first mapped to
a ConfigClass model and later generates an AS config class and
(optionally) and AS target class. All this is done in lazy fashion, and exactly once.