fixes JOO-82 EXML XSD generator fails in directories under a symlink http://jangaroo.myjetbrains.com/youtrack/issue/JOO-82
The bug is caused by checks whether a file found in the source
path is actually a generated file, which are implemented by
checking whether the source file is below the output folder.
These checks become obsolete with the right JangarooParser
configuration in ConfigClassRegistry: simply add the output
folder to classpath, not to sourcepath.
Adding the output folder to the sourcepath in ExmlXsdMojo
was already obsolete without that change: generated ActionScript
config classes never need to be scanned, because always the
EXML source is used.