New [Native] handling
Requiring the "native" classes and functions worked in most cases,
but it failed for package-scope variables. To support variables, too,
and to better support the dynamic nature of JavaScript (replace
classes or functions at runtime), we now require the package /
namespace (assuming it remains constant) and retrieve the class,
function or variable on every access.