fix: local variable shades class, even if qName is used
This happened because Jangaroo 3 always tries to use
the non-qualified name for imported classes, unless
there is a conflicting import.
Now, conflicts with other declarations are detected and
if present, the fully qualified name is used.
Note that the combination of a local variable shading
a top-level package AND a local variable shading the
class still does not work (hopefully an esoteric case).