Fix: private (non-static) const with expression initializer
Bug was they were not generated at all.
The fix is quite a hack, as the initializer expression is generated
into a string variable and then written into the "Jangaroo part"
of the code later. Better would be to handle such constants like
method declarations, where a local variable is defined with the
value exactly where the expression was located in the AS source
code, and then assign this variable to the corresponding property
in the "Jangaroo part".
This solution will still fail if the non-static const refers to a static
const that is no compile-time constant, because of a wrong
initialization order. To fix this, we would also need to handle static
const value assignment and access differently...