Reintroduce Maven goal war-package and simplify bootstrap
The goal is similar to generate-amd, but additionally
generates a module "jangaroo-application" which
requires the Maven AMD module of the web-app.
Thus, as in Jangaroo 1 + 2, the index.html can always
require "jangaroo-application" and does not have to
mention its groupId + artifactId.
Additionally, "jangaroo-application" is now also the
default module to load in the RequireJS plugin "run",
so to bootstrap a Jangaroo application you now simply
have to use "amd/run!" plus the fully qualified name of your
main class:

<script src="requirejs/require.js" data-main="amd/run!com.acme.MyMainClass"></script>