Only include default language properties in *.lib.js.
This resembles Jangaroo 2 behavior and leads to less data being
loaded to the client when only the default language is needed
(usually English) or when an application has many resource
bundles and/or languages.
Later, we should build locale-specific libraries, but rather we
should deprecate the EXML properties compiler in favor
of the new Flex-compatible properties compiler, and support
language libraries for that one.