Restore joo.baseUrl and IE style sheet limit workaround.
The logic used to determine joo.baseUrl has been restored, but
changed slightly, because instead of joo/jangaroo-application.js,
it now tries to find the requirejs/require*.js script.

In IE9, the 31 style sheets limit still exists, so we need to keep
the workaround. However, since style sheet URLs are collected
anyway, and since the solution for IE9 also works in all other
browsers, the code could be simplified a lot.

On the fly, reused more Jangaroo 2 code in jangaroo-runtime.js,
in order to closely match the old semantics. Among others, this
restores joo.resolveUrl().
Also, do not use console.log(), as it fails in IE9 when Developer
Tools are not open!