implements JOO-87 Allow to specify value of <exml:var> and <exml:constant> in a sub-element http://jangaroo.myjetbrains.com/youtrack/issue/JOO-87
This changes EXML's XSD to allow <exml:value> inside <exml:var> and <exml:constant>,
and to use any mixed content inside <exml:value>.
The value attribute of <exml:var> and <exml:constant> is no longer required, since it
has to be left out when using the sub-element.
Since this is a backwards-compatible change, the XSD URI does not change.

Note that nesting EXML elements in <exml:value> is currently only supported for <exml:var>,
not <exml:constant>. This is because the code that specifies the constant value is generated into
the config class, but we can only generate code from EXML elements into the target class.
Also, we do not see a use case where defining a constant value by EXML would be needed.

IDEA support for this new feature will be provided by an upcoming plugin update.