Updated JS-AS name mapping to avoid duplicate class names.
Though it is possible to have the same class name in different packages,
it is inconvenient, because it leads to using fully-qualified names.
Name clashes with top-level (built-in) classes must be avoided, because
when importing the qualified class, access to the top-level class is
no longer possible.

Some duplicate classes remain (for the time being) for compatibility
with Ext AS 3.4. When we have a migration tool, this decision might
be changed to rather use the new names and use the migration tool
to map the Ext 3.4 names.