Unique config class names for all Ext classes that support a config object.
So far not used by code, but be patient.

The config class name is usually the (x)type, but many Ext classes that
feature a constructor with a config object do not have an appropriate alias.
Also, new in Ext JS 6, aliases have a "namespace" prefix like "widget." or
"layout.", which leads to name clashes when using the name without the
namespace.