Introduce MXML constructor with config parameter.
Try 1: Using RegEx to parse ActionScript :-(

While original MXML only ever has a default constructor (a no-arg constructor),
EXML generates constructor with a single "config" parameter.
In order not to have to change ActionScript code that instantiates classes
formerly written in EXML using a "config" argument, we allow MXML classes
to declare a custom constructor, like so:

  public native function MyClass(config:myConfig = null);

The constructor must be native, because is body is generated from the other
MXML code. The parameter is coupled with an MXML element with the corresponding id:

  <cfg:myConfig id="config" .../>

Thus, all MXML expressions may reference "config" and resolve the correct config type.

The new MXML constructor is generated by ExmlToMxml and the changed format is
understood by MxmlToModelParser.

On the fly, to further reduce the diff between EXML and MXML, corrected white-space
in ASDoc comments and optimized order of top-level attributes.
(cherry picked from commit a53312d)