Generate EXML-compatible code for MXML code migrated from <exml:var>.
Disclaimer: MxmlToModelParser still uses RegEx to parse ActionScript :-(
But at least, only constructor and private field declarations have to be detected via RegEx.
We now got rid of exml:access="private" in favor of defining a private field in
ActionScript (inside <fx:Script>) and a top-level MXML element with the id of that
field's name.
Similarly, the actual name of the "config" constructor parameter is now also
parsed from the ActionScript constructor and the MXML element with the corresponding
id is used for config defaults.
Defaults are now applied before any vars are assigned, like it used to be in EXML.
The only thing that work slightly differently is that in EXML, the order of mixed <exml:var>s
with expression values and with XML values would be kept, while in MXML, you
put the vars with expression values into the constructor, while the vars with XML values are
always generated *behind* these. So you can no longer use an XML-valued var to
compute another var in an expression.