Made destination directory of catalog.xml configurable.
The jangaroo-maven-plugin now generates catalog.xml into the
output root directory, so that it is located in the root path of the
JAR artifact so that IDEA finds it. For IDEA to also accept JAR files
(not only SWC), our pull request to the Flex Plugin has to be accepted,
or a custom Flex Plugin has to be used.