Changed MXML initialisation syntax of former <exml:var>.
First approach was to give the usually native constructor a body like so:

  public function MyComponent(config:myConfig = null) {
    myVar = /* very complicated expression using "this" and "config" */;
    super(config);
  }

However, the actually generated constructor looked different: it added code
at the beginning of the constructor body, before the super call (initialization code
of  other vars declared in MXML) and replaced the super call parameter.

To keep all the magic in a function declared as native, we now always use a
native constructor and place custom initialization code in a private method with
the magic name "__initialize__", receiving the config object as its sole parameter.
The method is called by the generated constructor body after applying the config
defaults and before executing the MXML vars initialization.

ExmlToMxml and MxmlToModelParser have been adapted to this new semantics,
all tests are still green.