Re-enable mapping of immediately-created classes to be mapped to their target class.
This mapping is currently only used by IDEA, so that MXML elements like
<ui:someAction> refer to the target type SomeAction and fit into the corresponding
properties like baseAction:Action.
Code will still be red in IDEA, because the target classes do not yet contain the config
properties, so all properties are red. It is a manual, but straight-forward task to copy
all config properties from the config class to the corresponding target class and adjust
target class code if it uses private fields that clash with config properties.