For immediately-created classes, the target class is now added to the component library.
Since the last approach needed too many AS code changes to end up with green
properties in IDEA, we returned to the solution where immediate creation is made
explicit in MXML code. The disadvantage is that this makes the diff to EXML larger,
but this is outweighed by the advantages: there is less magic and no AS code changes.
The code is both green in IDEA and will compile with the same semantics as the original
EXML code.

This commit changes the ExmlToMxml converter, upcoming are the necessary changes
to the MXML compiler.