Teach MXML compiler to handle non-config classes and <fx:Declarations>...
...as now generated by ExmlToMxml.
So far, the target class is actually used only as a "type cast". The target object is
still instantiated from the config object using create(). This is essential if only
the config class is Public API, so that MXML usage of the target class does not
introduce new Public API Violations.