Ext JS 6 run-time compatibility.
Ext 6 requires super constructor calls to be made using the "superclass" property.
In order not to have to generate two different types of super calls, the Jangaroo
Runtime now also uses "superclass", which is a property of the constructor function
referencing the prototype (!) of the superclass.
Code generation is not yet perfect: A "package aux var" should be used, not the
fully-qualified name, but it seems it is not always available...