MXML compiler: parse variable declarations from script code.
If a variable is already declared in script code, an <fx:Declarations> element
using the same id does not generate an additional field, but just initializes the
existing one.
For public fields, the super config object is set, so that the value is eventually applied
to "this". For all other visibilities, the field is set directly.