To be SenchaCmd-compatible, use Ext.define() everywhere.
For non-Ext-classes, instead of using Ext.ClassManager.triggerCreated(), which is
not Ext public API and does not help to make custom scripts work with SenchaCmd,
wrap custom objects we want to require in an Ext.define().
The trick is to use singleton: true with a constructor that returns our custom object.
Note that this tricks fails for "falsy" values like null, 0, "", etc., so it cannot be used
for ActionScript package-scope variables. :-(