Changed strategy when to use config objects.
To be more compatible with EXML, we switch back to the EXML strategy
that decides between instantiating objects and using config objects (type casts).
In EXML, this did not depend on the property, but only on the used class.
In MXML, to force using config objects, we now look for the annotation
[ExtConfig] at the class and all its super classes. In ext-as, the classes
ext.Component, ext.plugin.AbstractPlugin, and ext.layout.Layout have
to be annotated.
Also, any class implementing ext.Plugin will use config objects. This is hard-wired
to avoid having to scan all interfaces, too.
The annotation [AllowContructorParameters] is kept, but with a reversed default:
it allows to specify properties that allow only pre-instantiated objects, no config objects.