Always let type cast add xclass to config object and complement xtype / type.
Now, a type cast of a primitive object always sets the xclass property, and sets
the xtype or type property if the class has the corresponding (x)type.
This keeps Ext happy when it fails to look for xclass, and at the same time avoids
situations where a wrong xclass wins over the correct (x)type.