Lazy class initialization closer to ActionScript 3 / Jangaroo 2.
Instead of executing static class code (static code plus static field initializers)
when Ext.define() calls the class-created-callback, we now intercept all access
to the constructor and all static members with a function that cleans up the
interceptors, executes the static class code, and eventually delegates to the
original constructor / static member.
This leads to static code being executed in a very similar way as in original
ActionScript 3 as well as Jangaroo 2.