Replace [Native] by [Mixin] for mixin interfaces.
Annotating a mixin interface with [Native] and the native name of the
implementation class was rather a hack and is not a clean solution when
we actually want to implement mixins in ActionScript/MXML.
Rather, the mixin interface should reference its mixin implementation class
with a dedicated annotation [Mixin] using its ActionScript name.
Then, the implementation class can be [Native] as well as "real" ActionScript/MXML.

Todo: Implement [Mixin] handling in JsCodeGenerator.