Allow implementing Ext Mixins in ActionScript and MXML
In ActionScript, an Ext Mixin is specified as a pair of interface
and implementation. The interface has to be annotated as
  [Mixin("<fully-qualified-name-of-implementation-class>"]

As before, any class that implements the Mixin interface (with the
exception of the Mixin implementation class itself) actually mixes
in the implementation class. To satisfy the standard ActionScript
type system (and thus IDEA), all interface methods have to be
declared, but simply as "native" (no method body). The Jangaroo
IDEA Plugin adjusts the "implement interface methods" action
(Ctrl-I) accordingly.