Fix super call to other super class method.
The check that the identifier is a method of the current class
it not only unnecessary, but harmful when the super call uses
a method that is not overridden in this class. While this does
not really make sense (you could just use this.foo()), it is allowed
and should work (it does now).