Do not redeclare "arguments" as local variable.
To convert the implicitly defined "arguments" variable to a "real" Array,
Jangaroo used to generate this code:
  var arguments = Array.prototype.slice.call(arguments);
While this worked at runtime, this local redeclaration of "arguments" lets
some JavaScript minifiers assume that "arguments" is only a local variable,
which is unused and consequently eliminated.
Now, Jangaroo leaves out the "var" keyword when converting "arguments"
to an Array.