Improved white-space handling for inserted code.
When the code generator inserts code like AS3.setBindable(...), it
used to do so before the next symbol. Since symbols have leading
white-space, that white-space was printed between the inserted
code and the symbol, leading to generated code like
  AS3.setBindable(
  this,"foo", "bar");AS3.setBindable(
  this,"shaz", "baz");

While this resembles the original source code, it makes debugging
harder, as statements start on the previous line.

Now, JsWriter takes care of white-space of each symbol to be written
at most once, so when symbol white-space is written in advance,
it is suppressed when the symbol is written.
This allows to write symbol white-space, then inserted code, then
the symbol, resulting in the desired code

  AS3.setBindable(this,"foo", "bar");
  AS3.setBindable(this,"shaz", "baz");