Fix: allow boolean constants in text content.
Using a boolean constant as text content of a boolean field like so
  <fx:Boolean id="foo">true</fx:Boolean>
(without curly braces) generated this undesired code:
   var foo:Boolean = new Boolean();
The text content was completely ignored and a Boolean (object) was
instantiated, which in JavaScript behaves oddly at run-time.

Now, boolean fields and properties allow the constants "true" and "false"
without curly braces. (Actually, they treat any text content that is
not a binding expression as code, but most of it will be red in IDEA or
lead to compiler errors later.)

While at it, improved MXML compiler to stop with a compiler error when
encountering unexpected text content in an MXML node.
Note that multiple text nodes, e.g. between/after array sub-elements, are
not yet detected.