Improve event names and classes (types).
Like xtypes, event names are now converted to CamelCase.
To restore words, here is a small "dictionary" of all words used
in Ext JS event names. The longest matching prefix contained in
event-words.properties is used as the next word, which works
for all existing event names. Whenever a new Ext release is used,
this list may have to be complemented.

Event classes are now reused for all events that use the same
callback method signature / event properties. The name of the
event class is generated from the issuing class name plus the
names of all additional event properties. While there is no
guarantee that this results in a unique class name, in practice,
there is no name-clash (at least not with current Ext JS 6.0.1).

Reusing event classes also means that the event name constant
is no longer called NAME, but is called like the event name and
defined in the class that represents the event type.

All event classes now inherit from a custom event base class
called net.jangaroo.ext.FlExtEvent, defined in ext-as. To make
things work in the IDE, this class might have to inherit from
flash.events.Event, or must be declared an event type itself.