Improve error message when accessor is not implemented.
When an interface declares get and set accessors for some property "x",
and an implementing class fails to implement both of them, the compiler
used to complain about "x" not being implemented. It was up to the
programmer to find out that while get x() was implemented, set x() was
missing, or the other way round.
Now the accessor type is added to the error message.

I stumbled across this situation because of an IDEA ActionScript bug,
which results in interface methods documented as @private (as set
accessors are supposed to be documented!) being ignored, so that
the implementing class was green in IDEA, but did not compile.