Support "type" (not only "xtype") in config objects.
Whenever the type you cast to has an "alias", in addition
to "xclass", a "type" property is now added to the config
object that stips the prefix from the alias.
For example, if you cast to the Proxy subclass ext.data.proxy.AjaxProxy
which defines the alias "proxy.ajax", the resulting config
object would look like so:

    {
      xclass: "Ext.data.proxy.Ajax",
      type: "ajax"
    }

Rationale: There is still a lot of Ext code that does not respect
"xclass", but looks for "<myprefix>." + config.type instead.
Proxy classes are an example that now works with MXML.