Expect Resource(bundle='...')] without _properties suffix.
Because we used the Flex annotation syntax, we should also use
the bundle name Flex uses to reference resource bundles, and
the _properties class suffix was just an EXML invention.

On the fly, allow bundle and key parameter in any sequence,
and allow both types of quotes to surround their values.

Note: All Resource() references in properties files have to be
stripped off their bundle's _properties suffix before switching
to this compiler version!