Remove common SenchaHelper interface.
Where is the benefit of helper classes implementing a common interface?
I guess it was supposed to enforce some kind of lifecycle.
Especially for SenchaWorkspaceHelper which does not even implement
packageModule(), this does not really make sense.
All three helper classes are only used in exactly one place, using the concrete
class, so the interface offers no abstraction.