Fix production vs. development app build, even for multiple locales.
We now call SenchaCmd with an explicit build environment,
'production' or 'development'. This can be chosen by the
system property senchaAppBuild, i.e.
  -DsenchaAppBuild=development
results in a development (only) build. Note that when you
do 'mvn install', an incomplete artifact will be installed.
The default is 'production'.

For the default locale, like before, a full build is executed.
For each additional locale, calling 'app refresh' did not suffice.
We now call 'app build', but skip sass and resources.
For 'development' builds, this is quite fast (no minification).