Compile Jangaroo sources directly to Sencha output directory.
Since unit tests are now run with SenchaCmd, the generated JavaScripts
are no longer needed in the JAR. Thus, they can be compiled directly
to their final destination in the target Sencha output directory.
For this to work with apps, too, I had to split app generation to two
phases: first, an empty app is generated and the example "app" folder
is deleted. Then, Jangaroo sources are compiled into that "app" folder.
Then, the application is prepared and packaged.