Fix properties source inclusion scanning.
We need a custom SourceMapping that is aware of how
properties source file names are mapped to JavaScript
output files. This needed a bit refactoring of PropcHelper,
which is still not ideal.

On the fly, fixed that input suffix filtering was done through
the SuffixMapping when no includes or excludes were given
(the usual case!), because then the default includes to use the
input file suffix was not applied.