Fix incorrect cyclic dependency detection.
In certain situations, the compiler would stop with an error about a cyclic static
dependencies that was not correct. This happened when only a subset of the
classes in the source path was handed in as source files. Files that are contained
in the source path but not in the source files are only scoped, not analyzed.
For detecting top-level static methods, DependencyGraph checked whether the
class body equals the static method's parent node, but the parent node had not
yet been set (it is set during analyze). Thus, the static method incorrectly lead to
an "init" dependency instead of a "static" dependency, leading to an incorrect
dependency cycle error. Thanks okummer for this analysis.

The fix is to not check the parent node at all. In ActionScript, static functions can
only occur directly inside a class body, anything else would be a compile error.
On the fly, changed the currentDeclaration from a Dependency array to simply a
Dependency. I guess StaticDependencyVisitor started as an inner class, where it
is necessary to wrap outer variables in an array to allow write access. For a separate
class, this is no longer necessary.