Introducing SkipJangarooAnalyzer
The dependencies in the Jangaroo module's POM are declared as JAR packages and as such the maven-dependency-plugin tries to analyse usages of the Java classes of the dependencies. As there are no Java classes inside these JAR dependencies you either have to force "usedDepencencies" by means of the maven-dependency-plugin in the Jangaroo module or you can use this analyzer which reports all Jangaroo JAR dependencies as being used.