Support for custom config get / set accessors.
For a config foo, Ext JS generates the methods getFoo() and setFoo().
However, in ActionScript, a config is represented by a [Bindable] property,
either implemented as a public field (var, already supported) or -- as
supported as of now -- a get / set accessor function pair.
Just like read and write of a config property are rewritten to use
AS3.get/setBindable() at run-time, overriding the get foo / set foo
accessors is now transformed into overriding the generic getFoo() / setFoo()
methods. This completes hiding the existence of these methods to the
Flex developer.

Caveat: Since such custom get / set config accessors use a private field in
Jangaroo-style (this._foo$1) instead of the usual Ext JS way (at least for classic
toolkit) to use the plain name (this.foo), simply applying the config object on
the target object won't work, but initConfig() has to be used, which again calls
the setter. Fortunately, this is the case anyway for all Component subclasses
and several other classes (see Ext documentation).