Introduce ExpressionType for more accurate typing.
This allows to distinguish three meta-types of types:
instance, class, and function.
An instance is the "normal" type, an instance of a (sub)class or a class
implementing a certain interface.
A class can be instantiated ("new"), used as a type cast (ApplyExpr), used
for access to static members (DotExpr, ToDo) or used as a Class instance (ToDo).
A function can be applied, resulting in an expression of its return type,
or used as a Function instance.

Now, many more expression types receive a type, resulting in much better
error detection and code rewrite detection (private members, bindables).
All tests, including JooUnit-self-tests, are green.