Use [ArrayElementType] to resolve ArrayIndexExpr type.
See TestEventListener.as for an example of where now an array element type can be detected
correctly, too.
While introducing new meta types ARRAY and VECTOR, refactored duplicate code from IdeExpr
and DotExpr into a new factory method in ExpressionType.

TODO:
When only get or set accessor is annotated, jooc does not look for the annotation at the
"opposite" element. But this is an open issue for [Bindable], too.