Fix unwanted lookup during scoping.
This was caused by createAuxVar() trying to find a "free" identifier to avoid name
clashes, even with inherited members. But in scope phase, calls to lookupDeclaration()
are dangerous, as other compilation units might not yet be scoped.
Changed createAuxVar() to simply check for used identifier names locally. While this
is not water-proof, it suffices for local aux vars (e.g. used in for each loops), because
such identifiers are only used by generated code, which is never located in a nested scope
but always in the same scope.