Workaround for get / set accessor confusion.
Bug: Compiler complains about override with non-matching signature
when compiling jooflash-core (jangaroo-libs).

Evaluation: The error reveals that actually, the signature of a get function
is compared to the signature of the set function in the superclass, which
of course is different.

Currently, CompilationUnits keep track of their members keyed by name.
The only distinction is between instance and static members.
This is correct with one exception: accessors may have two declarations
for the same name, namely a get and a set function. When looking for the
declaration of an identifier, only one declaration is returned.

Fix (workaround): For the time being, I disabled override signature checks for set functions
by not scoping them in the first place.

The correct fix would be to change declaration lookup to return a "virtual" declaration,
aggregating get and set. Then, the override check could pick the correct accessor
function.