Move new utility method introduced in e3a3c196.
Module jangaroo-utils is only for code shared between Maven- and IDEA-plugin.
When changing anything under jangaroo-tools-api, the IDEA-plugin has to be
updated. For that reason, utility methods only needed within the compiler should
remain in module jangaroo-compiler.
ExpressionType is the main client of the utility method isNumber(), so I moved it there.