Fix: ASDoc needs "event" parameter for [Bindable] on interface / native methods
When using [Bindable] on interface methods or native methods, the ASDoc tool (actually: the Flex compiler)
complains that this is not allowed.
According to http://googolflex.com/?p=138 this is a bug, and the post also contains a
workaround: just add an "event" parameter!
Since we currently do not use this parameter in Jangaroo, for the time being, we hard-wire
the "event" parameter to the value "DUMMY" in the generated API *.as file (joo-api/...).
Hopefully, this workaround can be removed after updating to a new ASDoc / Flex compiler
version (Apache).