[ExtAsApiGenerator] Improve adjusting to reference API
Before, a member was looked up in the reference API also in all superclasses.
This led to method signatures being "adjusted" in strange ways.
For example, ext.selection.SelectionModel#select() was adjusted to the select()
method of Ext AS 3.4's TreeSelectionModel.

Now, if the superclass has its own explicit mapping, member look-up is stopped.