WIP: Experimental: Get rid of (native) MXML constructor and config parameter / field.
Instead of the constructor, initConfig() is overridden and filled with generated code.
Instead of accessing properties of "config", directly access properties of "this".

ToDo:
We cannot set all properties after super.initConfig().
Some properties access other during initialization, which fails when no property has
been set. Example: "data" relies on "tpl" (not [Bindable]).
So it seems non-bindable properties still have to be set on the config object.
The plan is to do so for all properties that are initialized by an expression *not using "this"*.

(cherry picked from commit 3b50ec9)