Make ExmlToMxml consistent with changed MXML format
Although the ExmlToMxml tool to use lives on branch master (Jangaroo 2),
I wanted to keep this consistent with the recent MXML changes.
And of course the tests have to be green.

So the status is we still use "config", but it is no longer declared as a field,
but as a native function get.