Fix: Ignore white-space around binding expressions
Before, leading and trailing white-space would be concatenated with
the expression value, often leading to implicit toString() conversion
of the expression value object.
An MXML text node value was already trimmed before checking for
a binding expression in MxmlToModelParser#createValueCodeFromElement()
(line 382), but that only handled a special case.
Now, the more general MxmlUtils.valueToString() method does the
trimming. Note that this results in leading and trainling white-space
around {...} also being ignored in MXML attribute values like in
  title=" {computeTitle()} "
where now, both the leading and the trailing white-space is
removed and only the result of computeTitle() is used.
To keep this white-space, change this to
  title="{' ' + computeTitle() + ' '}"

This change is essential in combination with IDEA's code reformatting.
For complex binding expression, "Reformat code" introduces white-space
before and/or after the curly-braces expression and thus broke the code
with the more strict intepretation of the Jangaroo compiler. Now, this
white-space is ignored and everything continues to work.