Improve IOException error reporting
Catching IOExceptions inside the loop allows to map the exception to the current source file.

Also, the concrete subclass of the IOException is sometimes needed to understand the
exception "message", which e.g. for a java.nio.FileNotFoundException is just the file name.
Thus, added the exception's class name to the compiler error message.

Had to factor-out a method fileToSymbol() in JangarooParser to simplify logging an error
for a file when the location inside the file is unknown.