Allow to generate Sencha Workspace below an existing one
This should be allowed since there is a use case:
We have one big Sencha Workspace for the framework and the
application, making it easy to make changes to the framework
and debug them in the application immediately.
Then, the framework is released, and the application needs its
own Sencha Workspace, residing in a sub-directory of the
global Sencha Workspace.