Make JooUnit PhantomJS Resource Timeout Configurable
While the number of retries on PhantomJS crashes and resource timeouts
already was configurable, the actual resource timeout was hard-wired to
500 ms plus 100 ms for each retry, cut off at 1000 ms.
On slower machines and on Windows, even 1000 ms was not sufficient to
load ext-all.js, and even if 900 ms worked, it was annoying to have four
retries each time.
If all we want it to detect PhantomJS resource load deadlocks, it makes no
sense to increase that timeout on each retry, because it should be sufficiently
high to avoid false negatives (i.e. a long-lasting resource load mistaken as a
dead-lock) from the start.
Now, the default resource timeout is 10 seconds, which should still not hurt
because dead-locks should not be that frequent.
Additionally, the resource timeout can now be configured through the
property jooUnitResourceTimeout.