Optimize: Suppress ArrayElementType at Runtime
Added ArrayElementType to the list of annotations used only by the
compiler, not at runtime. This suppresses the metadata generated by
these annotations in the generated JS files and thus reduces their
size.
The constant for this annotation already existed, but had to be moved
to Jooc, where all the other annotation constants are located.