Fix Grammar to Keep All MXML Source Code
To be able to completely restore an MXML file from its AST,
scanner (flex) and parser (cup) had to be adapted to keep all
symbols.
Now, MxmlCompilationUnit knows its XML header and calling
toString() on that header plus its root node results in a
one-to-one copy of the MXML file.
This enables modifying the AST and then writing the file, which
is needed for migrating MXML declarations (upcoming).

All tests are green, Studio workspace compiles and still seems
to work as before.